+++ agents_plugins/robotmk.py
@@ -1077,6 +1077,7 @@ class RMKCtrl(RMKState, RMKPlugin):
 
                 for k in self.keys_to_encode:
                     if k in state:
+                        # Do not transfer HTML log if disabled in WATO
                         if k == 'htmllog' and self.global_dict['transmit_html'] == False:
                             state[k] = None
                         else:
@@ -1106,12 +1107,15 @@ class RMKCtrl(RMKState, RMKPlugin):
             # zlib bytestream is base64 wrapped to avoid nasty bytes wihtin the
             # agent output. The check has first to decode the base64 "shell"
             data_encoded = self.to_zlib(data)
+        elif encoding == 'utf_8':
+            # nothing to do, already in utf8
+            data_encoded = data
         else:
             # TODO: Catch the exception! (wrong encoding)!
             pass
         # as we are serializing the data to JSON, let's convert the bytestring
         # again back to UTF-8
-        return data_encoded.decode('utf-8')
+        return data_encoded.decode('utf-8')            
 
     def to_base64(self, data):
         data_base64 = base64.b64encode(data)