+++ CHANGELOG.md
@@ -5,6 +5,21 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Fixed 
+
+* Fixed two bugs in 1.6 bakery: double nested key in `robotmk.yml` for piggybackhost and robotdir (#141, #143)
+
+### Changed
+
+* log_rotation was stored as string in `robotmk.yml`
+
+### Added 
+
+* Added documentation for `robotmk.yml` format for non CEE users 
+
+
 ## 1.1.0 - 2021-06-14
 
 ### Changed